feat: allow deleting batches
Deleting a batch can now be done via `DELETE /scan/batch/:batchId`.
