+++ src/server.test.ts
@@ -1,21 +1,22 @@
 import { Application } from 'express'
 import request from 'supertest'
 import { buildApp } from './server'
-import { Scanner } from './scanner'
+import { Importer } from './importer'
 import election from '../election.json'
+import Store from './store'
 
-jest.mock('./scanner')
+jest.mock('./importer')
 
 let app: Application
-let scanner: Scanner
-let scannerMock: jest.Mocked<Scanner>
+let importer: Importer
+let importerMock: jest.Mocked<Importer>
 
-function makeMockScanner(): Scanner {
+function makeMockImporter(): Importer {
   return {
     addManualBallot: jest.fn(),
     configure: jest.fn(),
     doExport: jest.fn(),
-    doScan: jest.fn(),
+    doImport: jest.fn(),
     doZero: jest.fn(),
     getStatus: jest.fn(),
     unconfigure: jest.fn(),
@@ -23,21 +24,21 @@ function makeMockScanner(): Scanner {
 }
 
 beforeEach(async () => {
-  scanner = makeMockScanner()
-  scannerMock = scanner as jest.Mocked<Scanner>
-  app = buildApp(scanner)
+  importer = makeMockImporter()
+  importerMock = importer as jest.Mocked<Importer>
+  app = buildApp({ importer, store: await Store.memoryStore() })
 })
 
 test('GET /scan/status', async () => {
   const status = {
     batches: [],
   }
-  scannerMock.getStatus.mockResolvedValue(status)
+  importerMock.getStatus.mockResolvedValue(status)
   await request(app)
     .get('/scan/status')
     .set('Accept', 'application/json')
     .expect(200, status)
-  expect(scanner.getStatus).toBeCalled()
+  expect(importer.getStatus).toBeCalled()
 })
 
 test('POST /scan/configure', async () => {
@@ -47,7 +48,7 @@ test('POST /scan/configure', async () => {
     .set('Content-Type', 'application/json')
     .set('Accept', 'application/json')
     .expect(200, { status: 'ok' })
-  expect(scanner.configure).toBeCalledWith(
+  expect(importer.configure).toBeCalledWith(
     expect.objectContaining({
       title: '2020 General Election',
     })
@@ -55,16 +56,16 @@ test('POST /scan/configure', async () => {
 })
 
 test('POST /scan/scanBatch', async () => {
-  scannerMock.doScan.mockResolvedValue(undefined)
+  importerMock.doImport.mockResolvedValue(undefined)
   await request(app)
     .post('/scan/scanBatch')
     .set('Accept', 'application/json')
     .expect(200, { status: 'ok' })
-  expect(scanner.doScan).toBeCalled()
+  expect(importer.doImport).toBeCalled()
 })
 
 test('POST /scan/addManualBallot', async () => {
-  scannerMock.addManualBallot.mockResolvedValue(undefined)
+  importerMock.addManualBallot.mockResolvedValue(undefined)
   await request(app)
     .post('/scan/addManualBallot')
     .send({
@@ -72,7 +73,7 @@ test('POST /scan/addManualBallot', async () => {
     })
     .set('Accept', 'application/json')
     .expect(200, { status: 'ok' })
-  expect(scanner.addManualBallot).toBeCalled()
+  expect(importer.addManualBallot).toBeCalled()
 })
 
 test('POST /scan/invalidateBatch', async () => {
@@ -83,33 +84,33 @@ test('POST /scan/invalidateBatch', async () => {
 })
 
 test('POST /scan/export', async () => {
-  scannerMock.doExport.mockResolvedValue('')
+  importerMock.doExport.mockResolvedValue('')
 
   await request(app)
     .post('/scan/export')
     .set('Accept', 'application/json')
     .expect(200, '')
-  expect(scanner.doExport).toBeCalled()
+  expect(importer.doExport).toBeCalled()
 })
 
 test('POST /scan/zero', async () => {
-  scannerMock.doZero.mockResolvedValue()
+  importerMock.doZero.mockResolvedValue()
 
   await request(app)
     .post('/scan/zero')
     .set('Accept', 'application/json')
     .expect(200, { status: 'ok' })
-  expect(scanner.doZero).toBeCalled()
+  expect(importer.doZero).toBeCalled()
 })
 
 test('POST /scan/unconfigure', async () => {
-  scannerMock.unconfigure.mockResolvedValue()
+  importerMock.unconfigure.mockResolvedValue()
 
   await request(app)
     .post('/scan/unconfigure')
     .set('Accept', 'application/json')
     .expect(200, { status: 'ok' })
-  expect(scanner.unconfigure).toBeCalled()
+  expect(importer.unconfigure).toBeCalled()
 })
 
 test('GET /', async () => {
