+++ src/server.ts
@@ -1,16 +1,22 @@
 //
 // Just the HTTP glue to the functionality, no implementations.
-// All actual implementations are in scanner.ts
+// All actual implementations are in importer.ts and scanner.ts
 //
 
 import express, { Application, Request, Response } from 'express'
 import * as path from 'path'
 
-import SystemScanner, { Scanner } from './scanner'
+import SystemImporter, { Importer } from './importer'
 import Store from './store'
 import { Election } from './types'
+import { FujitsuScanner, Scanner } from './scanner'
 
-export function buildApp(scanner: Scanner): Application {
+export interface AppOptions {
+  store: Store
+  importer: Importer
+}
+
+export function buildApp({ store, importer }: AppOptions): Application {
   const app: Application = express()
 
   app.use(express.json())
@@ -18,16 +24,16 @@ export function buildApp(scanner: Scanner): Application {
   app.post('/scan/configure', (request: Request, response: Response) => {
     // store the election file
     const election = request.body as Election
-    scanner.configure(election)
+    importer.configure(election)
     response.json({ status: 'ok' })
   })
 
   app.post('/scan/scanBatch', async (_request, response) => {
     try {
-      await scanner.doScan()
+      await importer.doImport()
       response.json({ status: 'ok' })
     } catch (err) {
-      response.json({ status: `could not scan ${err}` })
+      response.json({ status: `could not scan ${err.message}` })
     }
   })
 
@@ -37,28 +43,36 @@ export function buildApp(scanner: Scanner): Application {
 
   app.post('/scan/addManualBallot', async (request, response) => {
     const { ballotString } = request.body
-    await scanner.addManualBallot(ballotString)
+    await importer.addManualBallot(ballotString)
     response.json({ status: 'ok' })
   })
 
   app.post('/scan/export', async (_request, response) => {
-    const cvrs = await scanner.doExport()
+    const cvrs = await importer.doExport()
     response.set('Content-Type', 'text/plain')
     response.send(cvrs)
   })
 
   app.get('/scan/status', async (_request, response) => {
-    const status = await scanner.getStatus()
+    const status = await importer.getStatus()
     response.json(status)
   })
 
+  app.delete('/scan/batch/:batchId', async (request, response) => {
+    if (await store.deleteBatch(request.params.batchId)) {
+      response.json({ status: 'ok' })
+    } else {
+      response.status(404).end()
+    }
+  })
+
   app.post('/scan/zero', async (_request, response) => {
-    await scanner.doZero()
+    await importer.doZero()
     response.json({ status: 'ok' })
   })
 
   app.post('/scan/unconfigure', async (_request, response) => {
-    await scanner.unconfigure()
+    await importer.unconfigure()
     response.json({ status: 'ok' })
   })
 
@@ -69,19 +83,21 @@ export function buildApp(scanner: Scanner): Application {
   return app
 }
 
-export interface Options {
+export interface StartOptions {
   port: number | string
   store: Store
-  scanner: SystemScanner
+  scanner: Scanner
+  importer: SystemImporter
   app: Application
 }
 
 export async function start({
   port = process.env.PORT || 3002,
   store = new Store(path.join(__dirname, '..', 'cvrs.db')),
-  scanner = new SystemScanner({ store }),
-  app = buildApp(scanner),
-}: Partial<Options> = {}) {
+  scanner = new FujitsuScanner(),
+  importer = new SystemImporter({ store, scanner }),
+  app = buildApp({ importer, store }),
+}: Partial<StartOptions> = {}) {
   await store.init()
   app.listen(port, () => {
     // eslint-disable-next-line no-console
